Fetch optix headers (#127)
* remove SlangSessionExtendedDesc and add compiler options to SlangDesc

* cleanup

* fetch optix

* pass optix header path to slang

* add findSharedLibraryPath

* match CUdeviceptr typedef with optix

* allow optix initialization to fail (raytracing not available)

* report skipped tests

* formatting

+++ src/cuda/cuda-api.h
@@ -13,7 +13,7 @@ void rhiCudaApiShutdown();
 using cuuint32_t = uint32_t;
 using cuuint64_t = uint64_t;
 
-using CUdeviceptr = uintptr_t;
+using CUdeviceptr = long long unsigned int;
 using CUdevice = int;
 using CUcontext = struct CUctx_st*;
 using CUmodule = struct CUmod_st*;
